environment-configs/requirements.txt
jsonschema~=4.19.0
pytest~=7.1.2
PyYAML~=6.0